Rename ocr referrals component with a generic name now that it can be used for species and communities.
